+++ src/TD.TDOffice/TD.TDOffice.Contracts/Entities/MCPartnerCenovnikItemEntity.cs
@@ -24,9 +24,9 @@ public class MCPartnerCenovnikItemEntity : ILSCoreEntity
     [NotMapped]
     public DateTime CreatedAt { get; set; }
     [NotMapped]
-    public int CreatedBy { get; set; }
+    public long CreatedBy { get; set; }
     [NotMapped]
-    public int? UpdatedBy { get; set; }
+    public long? UpdatedBy { get; set; }
     [NotMapped]
     public DateTime? UpdatedAt { get; set; }
 }
