+++ src/TD.TDOffice/TD.TDOffice.Contracts/Entities/MCPartnerCenovnikKatBrRobaIdEntity.cs
@@ -22,9 +22,9 @@ public class MCPartnerCenovnikKatBrRobaIdEntity : ILSCoreEntity
     [NotMapped]
     public DateTime CreatedAt { get; set; }
     [NotMapped]
-    public int CreatedBy { get; set; }
+    public long CreatedBy { get; set; }
     [NotMapped]
-    public int? UpdatedBy { get; set; }
+    public long? UpdatedBy { get; set; }
     [NotMapped]
     public DateTime? UpdatedAt { get; set; }
 }
