+++ src/TD.TDOffice/TD.TDOffice.Contracts/Entities/User.cs
@@ -30,9 +30,9 @@ public class User : ILSCoreEntity
     [NotMapped]
     public DateTime CreatedAt { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
     [NotMapped]
-    public int? UpdatedBy { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+    public long? UpdatedBy { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
     [NotMapped]
     public DateTime? UpdatedAt { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
     [NotMapped]
-    public int CreatedBy { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+    public long CreatedBy { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
 }
