+++ src/TD.Web/TD.Web.Admin/TD.Web.Admin.Contracts/Dtos/KomercijalnoWebProductLinks/KomercijalnoWebProductLinksGetDto.cs
@@ -4,5 +4,5 @@ public class KomercijalnoWebProductLinksGetDto
 {
     public long Id { get; set; }
     public int RobaId { get; set; }
-    public int WebId { get; set; }
+    public long WebId { get; set; }
 }
